+++ backend/src/main/java/com/pfa/medical_backend/dto/MedecinDTO.java
@@ -8,7 +8,7 @@ import lombok.Data;
 
 @Data
 public class MedecinDTO {
-    private Integer identifiantM;
+    private Long identifiantM;
     private String nomM;
     private String prenomM;
     private LocalDate dateNaissM;
